+++ nevergrad/parametrization/data.py
@@ -56,6 +56,35 @@ class BoundChecker:
         return True
 
 
+class Mutation(core.Parameter):
+    """Custom mutation or recombination
+    This is an experimental API
+
+    Either implement:
+    - `_apply_array`  which provides a new np.ndarray from a list of arrays
+    - `apply` which updates the first p.Array instance
+
+    Mutation should take only one p.Array instance as argument, while
+    Recombinations should take several
+    """
+
+    @property
+    def value(self) -> tp.Callable[[tp.Sequence["Array"]], None]:
+        return self.apply
+
+    @value.setter
+    def value(self, value: tp.Any) -> None:
+        raise RuntimeError("Mutation cannot be set.")
+
+    def apply(self, arrays: tp.Sequence["Array"]) -> None:
+        new_value = self._apply_array([a._value for a in arrays], arrays[0].random_state)
+        arrays[0]._value = new_value
+
+    def _apply_array(self, arrays: tp.Sequence[np.ndarray], rng: np.random.RandomState) -> np.ndarray:
+        raise RuntimeError("Mutation._apply_array should either be implementer or bypassed in Mutation.apply")
+        return np.array([])  # pylint: disable=unreachable
+
+
 # pylint: disable=too-many-arguments, too-many-instance-attributes
 class Array(core.Parameter):
     """Array parameter with customizable mutation and recombination.
@@ -235,7 +264,7 @@ class Array(core.Parameter):
                               "you should aim for at least 3 for better quality.")
         return self
 
-    def set_recombination(self: A, recombination: tp.Union[str, core.Parameter, utils.Crossover]) -> A:
+    def set_recombination(self: A, recombination: tp.Union[str, core.Parameter]) -> A:
         assert self._parameters is not None
         self._parameters._content["recombination"] = (recombination if isinstance(recombination, core.Parameter)
                                                       else core.Constant(recombination))
@@ -253,7 +282,7 @@ class Array(core.Parameter):
                 self.set_standardized_data(func(size=self.dimension), deterministic=False)
             else:
                 raise NotImplementedError('Mutation "{mutation}" is not implemented')
-        elif isinstance(mutation, utils.Mutation):
+        elif isinstance(mutation, Mutation):
             mutation.apply([self])
         else:
             raise TypeError("Mutation must be a string or a Mutation instance")
@@ -364,7 +393,7 @@ class Array(core.Parameter):
         if isinstance(recomb, str) and recomb == "average":
             all_arrays = [p.get_standardized_data(reference=self) for p in all_params]
             self.set_standardized_data(np.mean(all_arrays, axis=0), deterministic=False)
-        elif isinstance(recomb, utils.Crossover):
+        elif isinstance(recomb, Mutation):
             recomb.apply(all_params)
         else:
             raise ValueError(f'Unknown recombination "{recomb}"')
