+++ nevergrad/parametrization/test_utils.py
@@ -8,7 +8,6 @@ import time
 import contextlib
 import typing as tp
 from pathlib import Path
-import numpy as np
 from nevergrad.common import testing
 from . import parameter as p
 from . import utils
@@ -68,38 +67,6 @@ def test_flatten_parameter_order(order: int, keys: tp.Iterable[str]) -> None:
     param = p.Choice([p.Dict(x=p.Scalar(), y=12), p.Scalar().sigma.set_mutation(sigma=p.Scalar())])
     flat = helpers.flatten_parameter(param, with_containers=False, order=order)
     assert set(flat) == set(keys), f"Unexpected flattened parameter: {flat}"
-
-
-def test_crossover() -> None:
-    x1 = 4 * np.ones((2, 3))
-    x2 = 5 * np.ones((2, 3))
-    co = utils.Crossover(axis=1)
-    out = co._apply_array((x1, x2), rng=np.random.RandomState(12))
-    expected = np.ones((2, 1)).dot([[4, 5, 4]])
-    np.testing.assert_array_equal(out, expected)
-
-
-def test_rolling() -> None:
-    x = np.arange(4)[:, None].dot(np.ones((1, 2)))
-    roll = utils.Rolling(0)
-    out = roll._apply_array([x], rng=np.random.RandomState(12))
-    expected = np.array([1, 2, 3, 0])[:, None].dot(np.ones((1, 2)))
-    np.testing.assert_array_equal(out, expected)
-    assert repr(roll) == "Rolling(axis=(0,))"
-
-
-@testing.parametrized(
-    all_none=(None, None),
-    d2=((1, 2), None),
-    d1=((1), None),
-)
-def test_crossover_axis(axis: tp.Optional[tp.Tuple[int, ...]], max_size: tp.Optional[int]) -> None:
-    shape = (6, 8, 10)
-    x1 = 4 * np.ones(shape)
-    x2 = 5 * np.ones(shape)
-    co = utils.Crossover(axis=axis, max_size=max_size)
-    out = co._apply_array((x1, x2), rng=np.random.RandomState(12))
-    np.testing.assert_array_equal(out.shape, shape)  # this basically only test that it did not raise an error
 
 
 def test_descriptors() -> None:
