+++ nevergrad/parametrization/utils.py
@@ -10,7 +10,6 @@ import tempfile
 import subprocess
 import typing as tp
 from pathlib import Path
-import numpy as np
 from nevergrad.common import tools as ngtools
 
 
@@ -149,106 +148,3 @@ class CommandFunction:
                 subprocess_error = subprocess.CalledProcessError(retcode, process.args, output=stdout, stderr=stderr)
                 raise FailedJobError(stderr.decode()) from subprocess_error
         return stdout
-
-
-class Mutation:
-    """Custom mutation or recombination
-    This is an experimental API
-
-    Either implement:
-    - `_apply_array`  which provides a new np.ndarray from a list of arrays
-    - `apply` which updates the first p.Array instance
-
-    Mutation should take only one p.Array instance as argument, while
-    Recombinations should take several
-    """
-
-    def __repr__(self) -> str:
-        diff = ngtools.different_from_defaults(instance=self, check_mismatches=True)
-        params = ", ".join(f"{x}={y!r}" for x, y in sorted(diff.items()))
-        return f"{self.__class__.__name__}({params})"
-
-    def apply(self, arrays: tp.Sequence[tp.Any]) -> tp.Any:  # avoiding circular imports... restructuring needed eventually
-        new_value = self._apply_array([a._value for a in arrays], arrays[0].random_state)
-        arrays[0]._value = new_value
-
-    def _apply_array(self, arrays: tp.Sequence[np.ndarray], rng: np.random.RandomState) -> np.ndarray:
-        raise RuntimeError("Mutation._apply_array should either be implementer or bypassed in Mutation.apply")
-        return np.array([])  # pylint: disable=unreachable
-
-
-class Crossover(Mutation):
-    """Operator for merging part of an array into another one
-
-    Parameters
-    ----------
-    axis: None or int or tuple of ints
-        the axis (or axes) on which the merge will happen. This axis will be split into 3 parts: the first and last one will take
-        value from the first array, the middle one from the second array.
-    max_size: None or int
-        maximum size of the part taken from the second array. By default, this is at most around half the number of total elements of the
-        array to the power of 1/number of axis.
-
-
-    Notes
-    -----
-    - this is experimental, the API may evolve
-    - when using several axis, the size of the second array part is the same on each axis (aka a square in 2D, a cube in 3D, ...)
-
-    Examples:
-    ---------
-    - 2-dimensional array, with crossover on dimension 1:
-      0 1 0 0
-      0 1 0 0
-      0 1 0 0
-    - 2-dimensional array, with crossover on dimensions 0 and 1:
-      0 0 0 0
-      0 1 1 0
-      0 1 1 0
-    """
-
-    def __init__(self, axis: tp.Optional[tp.Union[int, tp.Iterable[int]]] = None, max_size: tp.Optional[int] = None) -> None:
-        self.axis = (axis,) if isinstance(axis, int) else tuple(axis) if axis is not None else None
-        self.max_size = max_size
-
-    def _apply_array(self, arrays: tp.Sequence[np.ndarray], rng: np.random.RandomState) -> np.ndarray:
-        # checks
-        arrays = list(arrays)
-        if len(arrays) != 2:
-            raise Exception("Crossover can only be applied between 2 individuals")
-        shape = arrays[0].shape
-        assert shape == arrays[1].shape, "Individuals should have the same shape"
-        # settings
-        axis = tuple(range(len(shape))) if self.axis is None else self.axis
-        max_size = int(((arrays[0].size + 1) / 2)**(1 / len(axis))) if self.max_size is None else self.max_size
-        max_size = min(max_size, *(shape[a] - 1 for a in axis))
-        size = 1 if max_size == 1 else rng.randint(1, max_size)
-        # slices
-        slices = []
-        for a, s in enumerate(shape):
-            if a in axis:
-                if s <= 1:
-                    raise ValueError("Cannot crossover an shape with size 1")
-                start = rng.randint(s - size)
-                slices.append(slice(start, start + size))
-            else:
-                slices.append(slice(0, s))
-        result = np.array(arrays[0], copy=True)
-        result[tuple(slices)] = arrays[1][tuple(slices)]
-        return result
-
-
-class Rolling(Mutation):
-
-    def __init__(self, axis: tp.Optional[tp.Union[int, tp.Iterable[int]]]):
-        self.axis = (axis,) if isinstance(axis, int) else tuple(axis) if axis is not None else None
-
-    def _apply_array(self, arrays: tp.Sequence[np.ndarray], rng: np.random.RandomState) -> np.ndarray:
-        arrays = list(arrays)
-        assert len(arrays) == 1
-        data = arrays[0]
-        if rng is None:
-            rng = np.random.RandomState()
-        axis = tuple(range(data.dim)) if self.axis is None else self.axis
-        shifts = [rng.randint(data.shape[a]) for a in axis]
-        return np.roll(data, shifts, axis=axis)  # type: ignore
